web/requirements.txt
alabaster==0.7.8
Babel==2.3.4
defusedxml==0.4.1
boto==2.39.0
cffi==1.8.3
coverage==4.2
cryptography==1.5.1
dj-database-url==0.4.0
Django==1.10.1
django-cors-middleware==1.3.1
django-filter==0.12.0
djangorestframework==3.4.6
docutils==0.12
gunicorn==19.4.5
imagesize==0.7.1
idna==2.1
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
Pillow==3.1.1
psycopg2==2.6.1
Pygments==2.1.3
pytz==2016.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.1.0
selenium==3.0.1
tblib==1.3.0
